+++ packages/components/src/components/FileUploader/FileUploader.test.tsx
@@ -127,23 +127,56 @@ describe("<FileUploader />", () => {
     });
   });
 
-  it('renders correctly when state is "error"', () => {
-    renderFileUploader({
-      label: "Visa Confirmation",
-      maxFileSize: "2MB",
-      fileName: "visa_v2.pdf",
-      fileUrl: FILE_URL,
-      fileSize: "1MB",
-      progress: 100,
-      errorMessage: "File is too large.",
+  describe("when there is an error", () => {
+    it("renders label and the Remove button if it has a fileName", () => {
+      renderFileUploader({
+        label: "Visa Confirmation",
+        maxFileSize: "2MB",
+        fileName: "visa_v2.pdf",
+        fileUrl: FILE_URL,
+        fileSize: "1MB",
+        progress: 100,
+        errorMessage: "File is too large.",
+      });
+
+      expect(screen.getByText("Visa Confirmation")).toBeInTheDocument();
+      expect(screen.getByText("visa_v2.pdf • 1MB")).toBeInTheDocument();
+      expect(
+        screen.getByRole("button", { name: "Remove file" })
+      ).toBeInTheDocument();
+      expect(screen.getByRole("alert")).toHaveTextContent("File is too large.");
     });
 
-    expect(screen.getByText("Visa Confirmation")).toBeInTheDocument();
-    expect(screen.getByText("visa_v2.pdf • 1MB")).toBeInTheDocument();
-    expect(
-      screen.getByRole("button", { name: "Remove file" })
-    ).toBeInTheDocument();
-    expect(screen.getByRole("alert")).toHaveTextContent("File is too large.");
+    it("renders Upload button but not Remove button when there is an errorMessage", () => {
+      renderFileUploader({
+        label: "Permit",
+        errorMessage: "You must upload a file",
+      });
+
+      expect(screen.getByText("Permit")).toBeInTheDocument();
+      expect(
+        screen.getByRole("button", { name: "Upload" })
+      ).toBeInTheDocument();
+      expect(
+        screen.queryByRole("button", { name: "Remove file" })
+      ).not.toBeInTheDocument();
+      expect(screen.getByRole("alert")).toHaveTextContent(
+        "You must upload a file"
+      );
+    });
+
+    it("renders Upload button but not alert if there is no errorMessage", () => {
+      renderFileUploader({
+        label: "Permit",
+        errorMessage: "",
+      });
+
+      expect(screen.getByText("Permit")).toBeInTheDocument();
+      expect(
+        screen.getByRole("button", { name: "Upload" })
+      ).toBeInTheDocument();
+      expect(screen.queryByRole("alert")).not.toBeInTheDocument();
+    });
   });
 
   it('renders correctly when state is "disabled"', () => {