+++ packages/components/src/components/FileUploader/FileUploader.tsx
@@ -70,6 +70,11 @@ const FileUploader = React.forwardRef<HTMLDivElement, FileUploaderProps>(
     const status = getStatus({ progress, fileUrl, errorMessage });
     const isMobileUploading = isMobile && status === "loading";
     const isUploadedWithoutName = status === "success" && !fileName;
+    const hasFile = fileName || fileUrl;
+    const shouldShowChildren =
+      status === "waiting" || (status === "error" && !hasFile);
+    const shouldShowRemoveButton =
+      (status === "error" && hasFile) || status === "success";
 
     return (
       <Box.div display="flex" flexDirection="column" gap="space25">
@@ -124,13 +129,13 @@ const FileUploader = React.forwardRef<HTMLDivElement, FileUploaderProps>(
           {isMobileUploading && (
             <FileUploaderProgressBar fileName={fileName} progress={progress} />
           )}
-          {status === "waiting" &&
+          {shouldShowChildren &&
             React.cloneElement(children, {
               fullWidth: isMobile,
               disabled,
               style: { alignSelf: "start" },
             })}
-          {(status === "error" || status === "success") && (
+          {shouldShowRemoveButton && (
             <RemoveButton disabled={disabled} onClick={onRemove} />
           )}
           {status === "loading" && <CancelUploadButton onClick={onCancel} />}