+++ packages/components/src/components/Icon/Icon.test.tsx
@@ -8,4 +8,13 @@ describe("<Icon />", () => {
     const renderedIcon = screen.getByTestId("icon");
     expect(renderedIcon).toBeInTheDocument();
   });
+
+  it("renders wrapper according 'as' property", () => {
+    render(
+      <Icon as="span" data-testid="icon" decorative icon="AcademicCapIcon" />
+    );
+
+    const renderedIcon = screen.getByTestId("icon");
+    expect(renderedIcon.tagName).toBe("SPAN");
+  });
 });