+++ packages/components/src/components/Icon/Icon.tsx
@@ -2,8 +2,10 @@ import React from "react";
 import * as HeroOutlineIcons from "@heroicons/react/24/outline";
 import type { SystemProp, Theme } from "@xstyled/styled-components";
 import { Box } from "../../primitives/Box";
-
-type IconNames = keyof typeof HeroOutlineIcons;
+import { IconName } from "./types/IconName";
+import { LucideIcons } from "./LucideIcons";
+import { isHeroIcon } from "./utils/isHeroIcon";
+import { isLucideIcon } from "./utils/isLucideIcon";
 
 export interface IconProps
   extends Omit<React.HTMLAttributes<HTMLDivElement>, "color"> {
@@ -14,11 +16,13 @@ export interface IconProps
   /** Changes the dispay style of the icon container. */
   display?: "block" | "flex" | "inline-block" | "inline-flex";
   /** Changes the icon being rendered. */
-  icon: IconNames;
+  icon: IconName;
   /** Adjusts the size of the icon being rendered. */
   size?: SystemProp<keyof Theme["sizes"], Theme>;
   /** Controls the aria-label of the icon rendered.  */
   title?: string;
+  /** Sets the render element of the component. */
+  as?: React.ComponentProps<typeof Box.div>["as"];
 }
 
 /** Icons are small graphical representation of a program or a function. */
@@ -31,6 +35,7 @@ const Icon = React.forwardRef<HTMLDivElement, IconProps>(
       icon,
       size = "sizeIcon40",
       title,
+      as = "div",
       ...props
     },
     ref
@@ -39,11 +44,18 @@ const Icon = React.forwardRef<HTMLDivElement, IconProps>(
       throw new Error(`${icon}: Missing a title for non-decorative icon.`);
     }
 
-    // eslint-disable-next-line import/namespace
-    const RenderedIcon = HeroOutlineIcons[icon];
+    let RenderedIcon = undefined;
+
+    if (isHeroIcon(icon)) {
+      // eslint-disable-next-line import/namespace
+      RenderedIcon = HeroOutlineIcons[icon];
+    } else if (isLucideIcon(icon)) {
+      RenderedIcon = LucideIcons[icon];
+    }
 
     return (
       <Box.div
+        as={as}
         color={color}
         display={display}
         h={size}
@@ -56,7 +68,9 @@ const Icon = React.forwardRef<HTMLDivElement, IconProps>(
           aria-label={title}
           as={RenderedIcon}
           display="block"
+          h={size}
           verticalAlign="middle"
+          w={size}
         />
       </Box.div>
     );