+++ packages/components/src/components/Icon/types/HeroIconName.ts
@@ -1,0 +1,3 @@
+import * as HeroOutlineIcons from "@heroicons/react/24/outline";
+
+export type HeroIconName = keyof typeof HeroOutlineIcons;