+++ packages/components/src/components/Icon/types/LucideIconName.ts
@@ -1,0 +1,9 @@
+export type LucideIconName =
+  | "baggage-claim"
+  | "building-2"
+  | "heart-handshake"
+  | "palette"
+  | "pie-chart"
+  | "puzzle"
+  | "slack"
+  | "wallet";