+++ packages/components/src/components/Icon/utils/isHeroIcon.ts
@@ -1,0 +1,8 @@
+import * as HeroOutlineIcons from "@heroicons/react/24/outline";
+import { IconName } from "../types/IconName";
+
+export const isHeroIcon = (
+  name: IconName
+): name is keyof typeof HeroOutlineIcons => {
+  return name in HeroOutlineIcons;
+};