+++ packages/components/src/components/Pagination/Pagination.stories.tsx
@@ -1,0 +1,65 @@
+import type { Meta, StoryObj } from "@storybook/react";
+
+import React from "react";
+import { Pagination } from "./Pagination";
+
+const meta: Meta<typeof Pagination> = {
+  title: "Components/Pagination",
+  component: Pagination,
+};
+
+export default meta;
+type Story = StoryObj<typeof Pagination>;
+
+const BasicPagination = (): JSX.Element => {
+  const [currentPage, setCurrentPage] = React.useState(5);
+  return (
+    <Pagination
+      currentPage={currentPage}
+      onPageChange={setCurrentPage}
+      pageCount={7}
+    />
+  );
+};
+
+export const Default: Story = {
+  render: (): JSX.Element => <BasicPagination />,
+};
+
+Default.parameters = {
+  docs: {
+    storyDescription:
+      "Pagination lets users navigate through content or a dataset that’s been broken up into multiple pages.",
+  },
+};
+
+const PaginationWithLotsOfPages = (): JSX.Element => {
+  const [currentPage, setCurrentPage] = React.useState(333);
+  return (
+    <Pagination
+      currentPage={currentPage}
+      onPageChange={setCurrentPage}
+      pageCount={999}
+    />
+  );
+};
+
+export const WithLotsOfPages: Story = {
+  render: (): JSX.Element => <PaginationWithLotsOfPages />,
+};
+
+const PaginationWithMorePageNeighbors = (): JSX.Element => {
+  const [currentPage, setCurrentPage] = React.useState(333);
+  return (
+    <Pagination
+      currentPage={currentPage}
+      onPageChange={setCurrentPage}
+      pageCount={999}
+      pageNeighbors={3}
+    />
+  );
+};
+
+export const WithMorePageNeighbors: Story = {
+  render: (): JSX.Element => <PaginationWithMorePageNeighbors />,
+};