+++ packages/components/src/components/Tabs/Tabs.stories.tsx
@@ -1,4 +1,4 @@
-import type { ComponentMeta } from "@storybook/react";
+import type { Meta } from "@storybook/react";
 import React from "react";
 import { Box } from "../../primitives/Box";
 import { Heading } from "../Heading";
@@ -9,13 +9,17 @@ import { TabPanels } from "./TabPanels";
 import { TabPanel } from "./TabPanel";
 import { Tab } from "./Tab";
 
-// eslint-disable-next-line @typescript-eslint/ban-ts-comment
-// @ts-ignore TabList has a required prop which confuses Storybook here.
-export default {
-  component: Tabs,
-  subcomponents: { Tab, TabList, TabPanels, TabPanel },
+const tabs: Meta<typeof Tabs> = {
   title: "Components/Tabs",
-} as ComponentMeta<typeof Tabs>;
+  args: {
+    // eslint-disable-next-line @typescript-eslint/ban-ts-comment
+    // @ts-ignore Table has a required prop which confuses Storybook here.
+    subcomponents: { Tab, TabList, TabPanels, TabPanel },
+    component: Tabs,
+  },
+};
+
+export default tabs;
 
 export const Default = (): JSX.Element => {
   return (
@@ -227,4 +231,48 @@ export const Fitted = (): JSX.Element => {
       </TabPanels>
     </Tabs>
   );
+};
+
+export const WithDivider = (): JSX.Element => {
+  return (
+    <Tabs>
+      <TabList aria-label="Page tabs" withDivider>
+        <Tab>Tab One</Tab>
+        <Tab>Tab Two</Tab>
+      </TabList>
+      <TabPanels>
+        <TabPanel>
+          <Box.div padding="space60">
+            <Heading as="h2" marginBottom="space0" size="heading50">
+              Tab One
+            </Heading>
+            <Paragraph>
+              Lorem ipsum dolor sit, consectetur. Aliquam ac a. Ligula at et,
+              sodales vel purus.
+            </Paragraph>
+          </Box.div>
+        </TabPanel>
+        <TabPanel>
+          <Box.div padding="space60">
+            <Heading as="h2" marginBottom="space0" size="heading50">
+              Tab Two
+            </Heading>
+            <Paragraph>
+              Lorem ipsum dolor sit, consectetur. Aliquam ac a. Ligula at et,
+              sodales vel purus.
+            </Paragraph>
+          </Box.div>
+        </TabPanel>
+      </TabPanels>
+    </Tabs>
+  );
+};
+
+WithDivider.parameters = {
+  docs: {
+    description: {
+      story:
+        "You can display a divider between the tabs and the content by using the `withDivider` property on the `TabList` component.",
+    },
+  },
 };